pyproject.toml
[project]
name = "alekhah"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit",
    "pypdf2",
    "langchain-google-genai>=3.0.3",
    "python-dotenv>=1.2.1",
    "typing-extensions>=4.15.0",
]
